requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
iso3166==2.1.1
iso3166-2==1.4.0
natsort==8.4.0
phonenumbers==8.13.29
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
typing_extensions==4.9.0
tortoise-orm==0.20.0
urllib3==2.2.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0